feat(migrations): add article properties to posts table
This commit adds two new columns, `is_hero_article` and `is_top_article`, to the `posts` table and the corresponding columns to the `_posts_v` view. It also creates indexes on these new columns to improve query performance.

The changes were made to support the addition of new article properties, which will be used to identify hero and top articles on the website.