feat(dependencies): update Payload CMS storage provider
The changes in this commit update the Payload CMS storage provider from `@payloadcms/storage-vercel-blob` to `@payloadcms/storage-s3`. This change is made to utilize the AWS S3 storage service instead of the Vercel Blob storage.

The key changes are:

- Update the dependency for `@payloadcms/storage-s3` to version `3.0.1`.
- Update the version information for the `@payloadcms/storage-s3` dependency to include additional AWS-related dependencies, such as `@aws-crypto/crc32`, `@aws-crypto/sha1-browser`, `@aws-crypto/sha256-browser`, and `@aws-sdk/client-s3`.

This change will allow the application to leverage the AWS S3 storage service for managing media and other files, providing a more robust and scalable storage solution.