feat(migrations): add image prefix column to media table
The changes in this commit add a new column `prefix` to the `media` table in the database. This column will store a default prefix for image URLs, which can be used to construct the full URL for an image.

The `up` function in the migration script adds the new column with a default value of `'musedam-website/media'`, while the `down` function removes the column if it exists.

This change is necessary to ensure that all image URLs in the application have a consistent prefix, which can be useful for managing and organizing media assets.