feat: Optimize blog post URLs
The changes made in this commit optimize the URLs for blog posts. Previously, the URLs were in the format `/blog/posts/:slug`, but now they have been simplified to `/blog/:slug`. This change makes the URLs more concise and easier to remember for users.

The main changes include:

- Update the `href` attribute in the `LocaleLink` components in `ArticleCard.tsx`, `HeroSection.tsx`, and `page.tsx` to use the new URL format.
- Update the `revalidatePost` function in `revalidatePost.ts` to use the new URL format when revalidating post paths.

These changes improve the overall user experience and make the blog section of the application more intuitive and user-friendly.