feat(payload-types): Add supported timezones and new payload-jobs type
This commit adds a new `SupportedTimezones` type to the `payload-types.ts` file, which
represents the list of supported timezones in IANA format. This type is referenced by
the `Config` interface's JSON-Schema via the `definition` property.

Additionally, this commit introduces a new `PayloadJob` type and corresponding
`PayloadJobsSelect` type to the `Config` interface. This allows for the management of
Payload jobs, which can be used for scheduling tasks such as publishing content.